feat: package the CLIs as exex.cli with console-script entry points, add CLI smoke tests
Move every scripts/*.py main() into src/exex/cli/<mod>.py exposing
main(argv=None) so the CLIs can run in-process; scripts/ become thin shims.
Add [project.scripts] entry points (exex-train, exex-eval, exex-router-stats,
exex-extract, exex-merge, exex-prune, exex-manage, exex-analyze, exex-report),
read the version dynamically from exex.__version__, promote datasets to a
required dependency, and export the cartridge/merger/pruner/loading/evaluate/
analyzer public names from exex.__all__.

tests/test_cli.py drives all nine CLIs against one saved tiny Gemma4 MoE
checkpoint per session (plus a tiny BPE tokenizer and a jsonl corpus),
removing large outputs as it goes so /tmp does not grow between runs.

Closes #34
Closes #36
