+++ scripts/extract_experts.py
@@ -1,48 +1,12 @@
 #!/usr/bin/env python3
-"""Extract experts from an MoE checkpoint into a cartridge file."""
+"""Thin shim; the implementation lives in ``exex.cli.extract`` (``exex-extract``)."""
 
-import argparse
-import json
 import os
 import sys
 
 sys.path.append(os.path.join(os.path.dirname(__file__), "..", "src"))
 
-import torch
-from transformers import AutoModelForCausalLM
-
-from exex.cartridge import save_cartridge
-
-
-def main():
-    parser = argparse.ArgumentParser(description=__doc__)
-    parser.add_argument("--model_path", required=True)
-    parser.add_argument("--experts", required=True,
-                        help='JSON dict {"name": index} or comma-separated indices')
-    parser.add_argument("--labels", default=None,
-                        help='JSON dict {"name": ["label", ...]}')
-    parser.add_argument("--output", required=True, help="Cartridge .safetensors path")
-    parser.add_argument("--dtype", default="bfloat16",
-                        choices=["bfloat16", "float16", "float32"],
-                        help="Load dtype; bf16 default matches native Gemma 4 "
-                             "checkpoints (fp16 casts lose precision / overflow)")
-    args = parser.parse_args()
-
-    try:
-        experts = json.loads(args.experts)
-    except json.JSONDecodeError:
-        experts = [int(i) for i in args.experts.split(",")]
-    labels = json.loads(args.labels) if args.labels else None
-
-    print(f"Loading model from {args.model_path}...")
-    model = AutoModelForCausalLM.from_pretrained(
-        args.model_path, torch_dtype=getattr(torch, args.dtype), device_map="cpu"
-    )
-    cart = save_cartridge(
-        model, experts, args.output, source_model=args.model_path, labels=labels
-    )
-    print(f"Wrote {args.output} with experts: {cart.expert_names}")
-
+from exex.cli.extract import main  # noqa: E402
 
 if __name__ == "__main__":
     main()
