+++ src/exex/cli/__init__.py
@@ -1,0 +1,6 @@
+"""Command-line entry points (``exex-*`` console scripts).
+
+Every module exposes ``main(argv=None)`` so the CLIs can be invoked
+in-process (tests) as well as from the ``scripts/`` shims and the
+``[project.scripts]`` entry points.
+"""
