+++ src/exex/cli/extract.py
@@ -1,0 +1,48 @@
+"""Extract experts from an MoE checkpoint into a cartridge file."""
+
+import argparse
+import json
+
+import torch
+from transformers import AutoModelForCausalLM
+
+from exex.cartridge import save_cartridge
+
+
+def build_parser():
+    parser = argparse.ArgumentParser(description=__doc__)
+    parser.add_argument("--model_path", required=True)
+    parser.add_argument("--experts", required=True,
+                        help='JSON dict {"name": index} or comma-separated indices')
+    parser.add_argument("--labels", default=None,
+                        help='JSON dict {"name": ["label", ...]}')
+    parser.add_argument("--output", required=True, help="Cartridge .safetensors path")
+    parser.add_argument("--dtype", default="bfloat16",
+                        choices=["bfloat16", "float16", "float32"],
+                        help="Load dtype; bf16 default matches native Gemma 4 "
+                             "checkpoints (fp16 casts lose precision / overflow)")
+    return parser
+
+
+def main(argv=None):
+    args = build_parser().parse_args(argv)
+
+    try:
+        experts = json.loads(args.experts)
+    except json.JSONDecodeError:
+        experts = [int(i) for i in args.experts.split(",")]
+    labels = json.loads(args.labels) if args.labels else None
+
+    print(f"Loading model from {args.model_path}...")
+    model = AutoModelForCausalLM.from_pretrained(
+        args.model_path, dtype=getattr(torch, args.dtype), device_map="cpu"
+    )
+    cart = save_cartridge(
+        model, experts, args.output, source_model=args.model_path, labels=labels
+    )
+    print(f"Wrote {args.output} with experts: {cart.expert_names}")
+    return cart
+
+
+if __name__ == "__main__":
+    main()
